pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0",
    "setuptools-scm>=8.0",
]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
]
name = "perovskite-solar-cell-database"
dynamic = ["version"]
description = "Perovskite solar cell data schema plugin for NOMAD."
readme = "README.md"
authors = [
    {name = "Jose Marquez", email = "jose.marquez@physik.hu-berlin.de"},
    {name = "Yaru Wang", email = "wangyaru@physik.hu-berlin.de"},
    {name = "Hampus Näsström", email = "hampus.naesstroem@physik.hu-berlin.de"},
    ]
maintainers = [
    {name = "Jose Marquez", email = "jose.marquez@physik.hu-berlin.de"},
    {name = "Hampus Näsström", email = "hampus.naesstroem@physik.hu-berlin.de"},
    ]
requires-python = ">=3.9"
dependencies = [
    "nomad-lab>=1.3.11.dev29",
    "nomad-schema-plugin-run>=1.0.1",
    "nomad-schema-plugin-simulation-workflow>=1.0.1",
    "rdkit",
    "openpyxl",
    "lxml_html_clean"
]
license = { file = "LICENSE" }

[project.urls]
"Homepage" = "https://github.com/FAIRmat-NFDI/nomad-perovskite-solar-cells-database"
Repository = "https://github.com/FAIRmat-NFDI/nomad-perovskite-solar-cells-database"

[tool.uv]
index-url = "https://gitlab.mpcdf.mpg.de/api/v4/projects/2187/packages/pypi/simple"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F401",  # Imported but unused
    "F403", # 'from module import *' used; unable to detect undefined names
    "E501", # Line too long ({width} > {limit} characters)
    "PLR2004", # Magic value used instead of constant
    "PLR0912", # Too many branches
    "PLR0915", # Too many statements
    "UP008",  # Use `super()` instead of `super(__class__, self)`
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]

[project.entry-points.'nomad.plugin']
perovskite_solar_cell = "perovskite_solar_cell_database:perovskite_solar_cell"
perovskite_solar_cell_database_app = "perovskite_solar_cell_database.apps:perovskite_solar_cell_database_app"
perovskite_composition = "perovskite_solar_cell_database:perovskite_composition"
ion_parser = "perovskite_solar_cell_database:ion_parser"
perovskite_ions_app = "perovskite_solar_cell_database.apps:perovskite_ions"
perovskite_tandem_cell = "perovskite_tandem_database:tandem_solar_cell"
solar_cell_app = "perovskite_solar_cell_database.apps:solar_cells"
